Add TDIGEST command info registration and update command metadata for CF.RESERVE and BF.ADD in Redis module; introduce new utility functions for documentation testing.
