+++ src/cmd_info/tdigest_info.c
@@ -1,0 +1,7 @@
+#include "redismodule.h"
+
+int RegisterTDigestCommandInfos(RedisModuleCtx *ctx) {
+    /* TODO: add TDIGEST command infos similar to CF. */
+    (void)ctx;
+    return REDISMODULE_OK;
+}
